requirements.txt
mazelib==0.9.16
playwright==1.44.0
jinja2==3.1.4
pytest==8.2.2
pytest-asyncio==0.23.7
numpy==1.26.4
Pillow>=10.0.0
scikit-image>=0.22.0
